[feat] modify : added some changes

+++ src/main/java/com/filmrental/model/entity/Address.java
@@ -12,7 +12,7 @@ public class Address {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "address_id")
-    private Short addressId;
+    private Integer addressId;
 
     @Column(name = "address", nullable = false, length = 50)
     private String address;
